Add null test
