--- Base32768.Test/Base32768StreamTest.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.IO;
-using FluentAssertions;
-using Xunit;
-
-namespace Kzrnm.Convert.Base32768
-{
-    public class Base32768StreamTest
-    {
-        [Fact]
-        public void NullArgumentConstructor_ReadMode()
-        {
-            Assert.Throws<ArgumentNullException>("reader", () => new Base32768Stream((TextReader)null));
-        }
-        [Fact]
-        public void NullArgumentConstructor_WriteMode()
-        {
-            Assert.Throws<ArgumentNullException>("writer", () => new Base32768Stream((TextWriter)null));
-        }
-        [Fact]
-        public void ReadSizeZero()
-        {
-            using var sr = new Base32768Stream(new StringReader("foo"));
-            sr.Read(new byte[0], 0, 0).Should().Be(0);
-        }
-    }
-}
